Implemented new idea for site
By putting a menu that would stick on top always, my idea here was to have under the sections and focus on showing my portfolio, which I consider important on a website like this.

I'm still implementing it, but I can say I like this project session with this styles!